modified by coding convetion and moved ScrollerResizableItem to Scroller.js
