--- packages/sampler/stories/qa-stories/components/ScrollerResizableItem/package.json
@@ -1,3 +1,0 @@
-{
-	"main": "ScrollerResizableItem.js"
-}
